invest 05 - added int to file length in samples

+++ dug_seis/acquisition/data_to_asdf.py
@@ -55,7 +55,7 @@ class DataToASDF:
         self._channel_count = len( param['Acquisition']['hardware_settings']['input_range'] )
         self._nr_of_data_points = floor(self.l_notify_size.value / 16 / 2)  # nr of channels & 16 bit = 2 bytes
         # self.file_length_in_samples = self._nr_of_data_points * 5  # a length that does not split transferred blocks
-        self.file_length_in_samples = self.file_length_sec * 1/self.stats['delta']
+        self.file_length_in_samples = int(self.file_length_sec * 1/self.stats['delta'])
         if self.file_length_in_samples < self._nr_of_data_points:
             logger.error('file_length_sec cannot be shorter than one buffer transfer: {} seconds'
                          .format(self._nr_of_data_points/self._sampling_rate))
